requirements.txt
asgiref==3.3.1
Django==3.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-summernote==0.8.11.6
django-taggit==1.1.0
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
